[refactor] add tester class
